moving code from test_compare_trajectory.py to test_ModelParameters.py
